backend/api/requirements.txt
Flask
Flask-Cors
python-dotenv

# Add all your original dependencies
torch>=2.2.0
numpy>=1.24.0,<2.0.0
transformers>=4.41.0
sentence-transformers>=2.2.2
pinecone>=6.0.0
anthropic>=0.18.0
scikit-learn
nltk
langchain-text-splitters